% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{add_stage}
\alias{add_stage}
\alias{add_stage.default}
\alias{add_stage.mongopipe}
\title{Create or append a mongo aggregation stage}
\usage{
add_stage(x, ...)

\method{add_stage}{default}(x, ...)

\method{add_stage}{mongopipe}(x, type, ...)
}
\arguments{
\item{x}{Object of class mongo or mongo_pipe}

\item{...}{query}

\item{type}{Name of the stage.}
}
\value{
Object of type mongopipe.
}
\description{
Create or append a mongo aggregation stage
}
