% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_operators.R
\name{condition}
\alias{condition}
\title{Conditional expression ($cond)}
\usage{
condition(test, yes, no)
}
\arguments{
\item{test}{Expression which returns a boolean value.}

\item{yes}{Return this if the test returns true.}

\item{no}{Return this if the test returns false.}
}
\value{
Return a list for using in mongopipe.
}
\description{
Evaluates a boolean expression to return one
of the two specified return expressions.
}
\examples{
\dontrun{
cond <- condition(test = list("$isArray"="$chart"),
                  yes = list("$size"="$chart"),
                  no = 0)
jsonlite::toJSON(cond)
}

}
