% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongopipe-package.R
\docType{package}
\name{mongopipe-package}
\alias{mongopipe-package}
\title{mongopipe: Write MongoDB Queries with R}
\description{
Translate R code into MongoDB aggregation pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpkgs.gitlab.io/mongopipe}
  \item \url{https://gitlab.com/rpkgs/mongopipe}
  \item Report bugs at \url{https://gitlab.com/rpkgs/mongopipe/-/issues}
}

}
\author{
\strong{Maintainer}: Oliver Haag \email{oliver_haag@e.mail.de}

}
\keyword{internal}
